Made sure to delete yesterdays jar file if we have been able to copy it
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@6175 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ nightly.py
@@ -641,7 +641,7 @@ if __name__ == '__main__':
         
     # get rid of  yesterdays distro jar file which we moved to nightly_dir
     # in anticipation of japi
-    os.unlink(oldCDKJar)
+    if oldCDKJar != None: os.unlink(oldCDKJar)
         
         
     # copy this script to the nightly we dir. The script should be in nightly_dir
